Feature analysis

See merge request wmols/ipsportdata!5

+++ domain/src/main/java/sogyo/wesley/ipsportdata/domain/IAnalysor.java
@@ -5,7 +5,7 @@ public interface IAnalysor {
 
     int getPower();
 
-    float getLactate_two();
+    double getLactate_two();
 
-    float getLactate_one();
+    double getLactate_one();
 }
